Added new epserde(phantom(...)) attribute

+++ examples/bench_vfilter.rs
@@ -125,8 +125,8 @@ where
         bench(args.n, args.repeats, || {
             let mut u = 0usize;
             for i in 0..args.n {
-                let s = packed_offsets[i] as usize;
-                let e = packed_offsets[i + 1] as usize;
+                let s = packed_offsets[i];
+                let e = packed_offsets[i + 1];
                 let q = unsafe { std::str::from_utf8_unchecked(&packed[s..e]) };
                 u ^= filter.contains(q) as usize;
             }
@@ -178,8 +178,8 @@ where
             bench(args.n, args.repeats, || {
                 let mut u = 0usize;
                 for i in 0..args.n {
-                    let s = packed_offsets[i] as usize;
-                    let e = packed_offsets[i + 1] as usize;
+                    let s = packed_offsets[i];
+                    let e = packed_offsets[i + 1];
                     let q = unsafe { std::str::from_utf8_unchecked(&packed[s..e]) };
                     u ^= filter.contains(q) as usize;
                 }
@@ -192,8 +192,8 @@ where
             bench(args.n, args.repeats, || {
                 let mut u = 0usize;
                 for i in 0..args.n {
-                    let s = packed_offsets[i] as usize;
-                    let e = packed_offsets[i + 1] as usize;
+                    let s = packed_offsets[i];
+                    let e = packed_offsets[i + 1];
                     let q = unsafe { std::str::from_utf8_unchecked(&packed[s..e]) };
                     u ^= filter.contains(q) as usize;
                 }
