+++ examples/bench_vfunc.rs
@@ -99,8 +99,8 @@ where
             bench(args.n, args.repeats, || {
                 let mut u = 0usize;
                 for i in 0..args.n {
-                    let s = packed_offsets[i] as usize;
-                    let e = packed_offsets[i + 1] as usize;
+                    let s = packed_offsets[i];
+                    let e = packed_offsets[i + 1];
                     let q = unsafe { std::str::from_utf8_unchecked(&packed[s..e]) };
                     u ^= func.get(q);
                 }
@@ -112,8 +112,8 @@ where
             bench(args.n, args.repeats, || {
                 let mut u = 0usize;
                 for i in 0..args.n {
-                    let s = packed_offsets[i] as usize;
-                    let e = packed_offsets[i + 1] as usize;
+                    let s = packed_offsets[i];
+                    let e = packed_offsets[i + 1];
                     let q = unsafe { std::str::from_utf8_unchecked(&packed[s..e]) };
                     u ^= func.get(q);
                 }
